% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasedcoin.R
\name{biasedcoin}
\alias{biasedcoin}
\title{Simulation of Biased Coin phase I design}
\usage{
biasedcoin(true_probs, start, q_esc, npat, nsim, seed)
}
\arguments{
\item{true_probs}{A vector of the true toxicity rates for each dose}

\item{start}{Starting dose}

\item{q_esc}{The biased coin's probability of "success"}

\item{npat}{Total number of patients available for the phase I study}

\item{nsim}{Number of simulations to run}

\item{seed}{If not empty, the seed to use for random generation}
}
\value{
The proportion of simulated trials that identified each dose as the MTD
}
\description{
The biasedcoin function simulates a biased coin phase I design. The function provides
the proportion of times each dose was chosen as the Maximum Tolerated Dose (MTD) and
the average Dose Limiting Toxicity (DLT) rate.
}
