% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storersA.R
\name{storersA}
\alias{storersA}
\title{Simulation of Storer's 3+3 phase I design}
\usage{
storersA(true_probs, nsim, seed)
}
\arguments{
\item{true_probs}{A vector of the true toxicity rates for each dose}

\item{nsim}{Number of simulations to run}

\item{seed}{If not empty, the seed to use for random generation}
}
\value{
The proportion of simulated trials that identified each dose as the MTD
}
\description{
The storersA function simulates Storer's phase I 3+3 design. The function provides
the proportion of times each dose was chosen as the Maximum Tolerated Dose (MTD).
}
