% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storersD.R
\name{storersD}
\alias{storersD}
\title{Simulation of Storer's 3+3 D phase I design}
\usage{
storersD(true_probs, npat, nsim, seed)
}
\arguments{
\item{true_probs}{A vector of the true toxicity rates for each dose}

\item{npat}{Total number of patients available for the phase I study}

\item{nsim}{Number of simulations to run}

\item{seed}{If not empty, the seed to use for random generation}
}
\value{
The proportion of simulated trials that identified each dose as the MTD
}
\description{
The storersD function simulates Storer's phase I D 3+3 design. The function provides
the proportion of times each dose was chosen as the Maximum Tolerated Dose (MTD).
}
